Added description for fixes
Added possiblestatesX variable because we didnt realise that emission matrix actually has rows that are number of states of observed variable X, and cols which are number of states of hidden variable Z.

Forward and backward parts are working, but if you have 4 states, you have to input numbers between 0 and 3 or program will not work.

That is purely programming problem, because we are only interested for the number of state, it doesn't matter what that state actually is.

Example:

We have 4 states, for 0 to 3. Each state stands for 1 letter of the alphabet, 0 for A, 1 for B...
Now, if we place actual letters for input, our program will not work, but if we say, first was letter A, which is state 0, and so on, program will work fine.

Again, shoutout to fantastic001 for making some of these thing clear.